Set MAVEN_OPTS to empty string if no proxy server
When there is no proxy server in use, assign the MAVEN_OPTS environment
variable to the empty string ("") instead of the space character
(" ") when exporting it.

Set the Maven phase to 'package' as there is no need to install the
results of the build into the local repository of the build machine.